requirements.txt
appdirs==1.4.3
Django==1.10.6
django-webpack-loader==0.4.1
djangorestframework==3.6.2
django-filter==1.0.2
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
six==1.10.0
virtualenv==15.1.0
